Migrate ParquetRsLibrary to @LibrarySpecification
